requirements.txt
pandas
requests
openpyxl
python-dotenv
openai
httpx
